Add new modules, fix all tests, and expand codebase
- Add 3D printing, VR editor, display editor, LSP text editor,
  node graph editor, weather editor, config schema, audio graph,
  LOD generator, PBR utils, render graph, shader graph, custom
  titlebar, asset dependency graph, transaction manager, and
  more core subsystems
- Add license plate generator, ERS hybrid system, Vulkan viewport
  item, AC event bridge, and server config editor
- Add tree-sitter syntax highlighting, minimap widget, script
  debugger, coroutine manager, hot reload, and debugger frontend
- Add GLTF, Alembic, USD parsers, format validator, audio format
- Add 8 new i18n translations (cn-tw, cn-zh, es, fr, pt-br, ru, tr)
- Add VR QML settings page, light theme QSS, translation script
- Update all test suites: add PhysicsSimulator tests (86 total),
  showroom editor test, and fix existing tests
- Update .gitignore to exclude temp files, moc outputs, and
  external tools (7zip, openxr)
- All 18 test suites pass: 262 tests, 0 failures

+++ src/modules/modellingEditor/TrackBuilder/cameratrackEditor/TrackCameraEditorModule.cpp
@@ -9,6 +9,8 @@
 #include <QSplitter>
 #include <QTextStream>
 #include <QFormLayout>
+#include <QJsonArray>
+#include <QJsonObject>
 
 namespace ks {
 
@@ -433,7 +435,7 @@ QJsonObject TrackCameraEditorModule::serializeProject() const
     for (const auto& cam : m_cameras) {
         QJsonObject obj;
         obj["id"] = cam.id;
-        obj["type"] = cam.type;
+        obj["type"] = static_cast<int>(cam.type);
         obj["name"] = cam.name;
         obj["fov"] = static_cast<double>(cam.fov);
         obj["nearClip"] = static_cast<double>(cam.nearClip);
